Move getYesterday into date.js file
